[Mouse Jump] - fix problem with some monitor layouts (#24607)
